requirements.txt
streamlit
pandas
numpy
matplotlib
seaborn
ydata-profiling
scikit-learn
python-docx
requests
xgboost